is_braces_constructible and supporting any_type in traits.h
Add tie_struct.h

tie_struct binds a structure as a tuple of references.

tuple_to_typelist extracts the types of a tuple to a typelist

member_typelist_t extracts the types of members of a struct as a typelist.

Added value_list to const_util.h

The other helpers in const_util don't yet work on value_lists.

Added soa_layout.h

soa_layout<T, Size> is a helper to calculate pointer offsets into a strided SoA
memory region of Size bytes. The T must be a structure of pointers. This can
be used to do efficient static SoA layouts for fixed-size chunks, similar to what
we do for ECS Chunks, but statically.