Split material loading to Material class
Added a temporary concept of Loader to render library.

It is a general interface for loading render resources. It needs to be converted to
async-only (with generators?), split some concepts between data definition types
and runtime types (e.g. material definitions vs instances) and support fast access
to previously-loaded types.