imgui has an unused variable when assertions are disabled
We could make the assertion condition still checked, but I'm distrustful of perf overhead there (really wish we had a [[pure]] attribute that could be compiler-checked...).

Until we add a way to set compiler flags only for Potato code that doesn't affect other dependencies (for -Werror and -Wall and the like) this will have to do.