Plumb through UUIDs in property grid as a second kind of asset reference
Note, this really isn't good. In fact, it's wrong in the case of selecting logical assets (different shader stages).

We need one kind of AssetRef that can reference a UUID+logicalName, which can be encoded as the full UUID+name in JSON and the condensed assetId in binary encodings.

Yes, this will break scene encodings.